Pong clone game
I used an example of a pong clone from free code camp. I made some changes from the original code. For example, in classic pong the "ball" is a square, so in the code I renamed the ball a puck. Pong is a good project for a programming student, many areas this code can be altered (change the player names, change the colors of the paddle, change the font of the text, etc)