Pong game clone
I learned how to make a pong clone from using the source code from free code camp and other online sources. I made some adjustments, such as renaming the players, and renaming the ball, a puck (since I always thought of pong, as a video game version of air hockey). This is a good project for programming students, the size of the screen and the paddles can be adjusted, the speed of the ball and speed of the paddles can be cahnged.