2 player
I may revisited this example if I ever want to expand on my card game.
This project has given me an appreciation of the work involved in expanding a program, to be more "complex"
When I first got into programming, I thought of coding in terms of , have an idea and keep coding into you can make that program the way you want. I think a better strategy is to start off small, and expand.